Connect to CORI/RISI Postgres DB

Connect to CORI/RISI Postgres DB

connect_to_db(
  schema,
  host = "cori-risi.c6zaibvi9wyg.us-east-1.rds.amazonaws.com",
  dbname = "data",
  port = 5432,
  config_name = NULL,
  config_file = "../base/config.yml"
)

Arguments

schema
A vector of schema names to add to the search path (e.g. 'sch_layer', sch_broadband', etc.)

host
Database host, defaults to the host for the main MDA Potgres database.

dbname
Database name to connect to. Defaults to main MDA Postgres database

port
Port name to connect to. Defaults to port for main MDA Postgres database

config_name
An optional name of a set of credentials in a config file. Must contain fields 'user' and 'password'. Alternatively, set the environment variables 'DB_USER' and 'DB_PWD' in R Studio.

config_file
Path to the config file to use, if not using environment variables. Defaults to "../base/config.yml"

Value

A pqconnection object