Copy a table from one schema to another within the same CORI/RISI Postgres DB instance ... will fail if table already exists in to_schema: ERROR: relation "..." already exists

Copy a table from one schema to another within the same CORI/RISI Postgres DB instance ... will fail if table already exists in to_schema: ERROR: relation "..." already exists

Usage

copy_table_across_schemas(
  from_schema,
  to_schema,
  table_name,
  grant_select_roles = c("read_only_access"),
  overwrite = FALSE
)

Arguments

from_schema

schema name to copy the target table from

to_schema

schema name to copy the target table to

table_name

name of the table that will be copied

grant_select_roles

vector of role names that will be granted permission to select table rows (e.g. c("read_only_access","r_team"))

overwrite

if TRUE and table exists in to_schema, it will be dropped and recreated (default = FALSE)

Value

the status of the copy transaction