Write data to the database. Optionally update metadata tables in the process

Write data to the database. Optionally update metadata tables in the process

Usage

write_db(
  con,
  name,
  dta,
  overwrite = FALSE,
  append = FALSE,
  spatial = FALSE,
  field_meta = NULL,
  table_meta = NULL,
  source_meta = NULL
)

Arguments

con

A database connection

name

Name of the table to write to

dta

A data frame or sf data frame

overwrite

If `TRUE`, overwrite an existing table

append

If `TRUE`, append to an existing table

spatial

If `TRUE`, write as spatial data

field_meta

An optional data frame matching the structure for field metadata generated by `create_metadata_tables()`

table_meta

An optional data frame matching the structure for table metadata generated by `create_metadata_tables()`

source_meta

An optional data frame matching the structure for source metadata generated by `create_metadata_tables()`